// src/app/model/user/user.ts
export class User {
  firstName: string;
  lastName: string;
  phone: string;
  mobile: string;
  isMobileVerify: boolean;
  email: string;
  isEmailVerify: boolean;
  password: string;
  passwordConfirm: string;
  role: string;
  // photo: Buffer;
  organizationalUnit: string;
  organizationalLevel: string;
  address: string;
  city: string;
  province: string;
  isActive: boolean;
}